trying to fix the eat() method, did fix the neural network crashing program
Some of this stuff is me trying to fix the eat function, so that the prey can actually eat the food, hasn't worked so far. But also the neural network no longer crashes the program. Wasn't because of the neural network itself, but cause of one while loop i put in which should be an if statement.